perf: Move keys usage mapping in interactive missing-keys fix outside of loop
The interactive fix of `missing-keys` (`add_missing_keys_interactively()`) previously called `map_keys_to_files()` inside a loop, once for every single missing key (to determine in which files it was used in) and therefore re-reading files for each key. Now it is only called once before the loop with all the missing keys.