feat: add nonexistent-keys --fix argument and functionality (#77)
* feat: add nonexistent-keys --fix argument and functionality

Add interactive fix for nonexistent-keys that allows users to interactively add values for nonexistent keys to the i18n source file.

* Always enforce sorted keys in i18n source file when adding new entries

* perf: Move keys usage mapping in interactive missing-keys fix outside of loop

The interactive fix of `missing-keys` (`add_missing_keys_interactively()`) previously called `map_keys_to_files()` inside a loop, once for every single missing key (to determine in which files it was used in) and therefore re-reading files for each key. Now it is only called once before the loop with all the missing keys.

* Update changelog and minor code refactoring

* Minor update to prompt for nonexistent keys fix

---------

Co-authored-by: Andrew Tavis McAllister <andrew.t.mcallister@gmail.com>